app/pyproject.toml
[project]
name = "pen-match-api-v2"
version = "0.1.0"
description = "PEN Match API V2 built with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings",
    "python-multipart",
    "openai",
    "azure-identity",
    "azure-core",
    "azure-cosmos==4.5.1",
    "azure-keyvault-secrets",
    "azure-storage-blob",
    "azure-search-documents",
    "azure-ai-documentintelligence",
    "beautifulsoup4",
    "structlog",
    "psutil",
    "httpx",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langgraph",
    "tavily-python",
    "redis",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"